refactor refreshViewers
