+++ bundles/org.eclipse.ui.ide/src/org/eclipse/ui/ide/fileSystem/zip/OpenZipFileHandler.java
@@ -19,22 +19,15 @@ import org.eclipse.core.resources.ZipFileTransformer;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.ISelection;
-import org.eclipse.jface.viewers.ISelectionProvider;
 import org.eclipse.jface.viewers.IStructuredSelection;
-import org.eclipse.jface.viewers.StructuredViewer;
 import org.eclipse.swt.widgets.Shell;
-import org.eclipse.ui.IViewReference;
-import org.eclipse.ui.IWorkbenchPage;
-import org.eclipse.ui.IWorkbenchPart;
-import org.eclipse.ui.IWorkbenchWindow;
-import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.handlers.HandlerUtil;
 import org.eclipse.ui.internal.ide.IDEWorkbenchPlugin;
 
 /**
  * This class represents a handler for opening zip files.
  *
- * @since 3.132
+ * @since 3.23
  */
 public class OpenZipFileHandler extends AbstractHandler {
 
@@ -48,54 +41,19 @@ public class OpenZipFileHandler extends AbstractHandler {
 	public Object execute(ExecutionEvent event) {
 		Shell shell = HandlerUtil.getActiveShell(event);
 		ISelection selection = HandlerUtil.getCurrentSelection(event);
-		if (!(selection instanceof IStructuredSelection)) {
+		if (!(selection instanceof IStructuredSelection structuredSelection)) {
 			return null;
 		}
-
-		Object element = ((IStructuredSelection) selection).getFirstElement();
-		if (!(element instanceof IFile)) {
+		if (!(structuredSelection.getFirstElement() instanceof IFile file)) {
 			return null;
 		}
-
 		try {
-			ZipFileTransformer.openZipFile((IFile) element, true);
-
-			if (shell != null && !shell.isDisposed()) {
-				shell.getDisplay().asyncExec(() -> {
-					refreshAllViewers();
-
-					// only for active viewer
-//					StructuredViewer viewer = (StructuredViewer) HandlerUtil.getActivePart(event).getSite()
-//							.getSelectionProvider();
-//					if (viewer != null && viewer.getControl() != null && !viewer.getControl().isDisposed()) {
-//						viewer.refresh();
-//					}
-				});
-			}
+			ZipFileTransformer.openZipFile(file, true);
+			ZipFileHandlerUtil.refreshAllViewers();
 		} catch (CoreException e) {
 			IDEWorkbenchPlugin.log(e.getMessage(), e);
 			MessageDialog.openError(shell, "Error opening zip file", e.getMessage()); //$NON-NLS-1$
 		}
 		return null;
-	}
-
-	private static void refreshAllViewers() {
-		IWorkbenchWindow window = PlatformUI.getWorkbench().getActiveWorkbenchWindow();
-		if (window != null) {
-			IWorkbenchPage page = window.getActivePage();
-			if (page != null) {
-				for (IViewReference viewReference : page.getViewReferences()) {
-					IWorkbenchPart part = viewReference.getPart(false);
-					if (part != null) {
-						ISelectionProvider selectionProvider = part.getSite().getSelectionProvider();
-						if (selectionProvider instanceof StructuredViewer viewer) {
-							if (viewer.getControl() != null && !viewer.getControl().isDisposed()) {
-								viewer.refresh();
-							}
-						}
-					}
-				}
-			}
-		}
 	}
 }
